+++ sample-apps/client/ts-quickstart/src/participant.ts
@@ -1,8 +1,4 @@
-import {
-  StreamVideoParticipant,
-  Call,
-  SfuModels,
-} from '@stream-io/video-client';
+import { Call, StreamVideoParticipant } from '@stream-io/video-client';
 
 // The quickstart uses fixed video dimensions for simplification
 const videoDimension = {
@@ -10,72 +6,98 @@ const videoDimension = {
   height: 250,
 };
 
-const renderVideo = (call: Call, participant: StreamVideoParticipant) => {
-  let videoEl = document.getElementById(
-    `video-${participant.sessionId}`,
-  ) as HTMLVideoElement | null;
+const videoBindingsCache = new Map<string, Function | undefined>();
+const videoTrackingCache = new Map<string, Function | undefined>();
+const audioBindingsCache = new Map<string, Function | undefined>();
 
+const renderVideo = (
+  call: Call,
+  participant: StreamVideoParticipant,
+  parentContainer: HTMLElement,
+) => {
+  const id = `video-${participant.sessionId}`;
+  let videoEl = document.getElementById(id) as HTMLVideoElement | null;
   if (!videoEl) {
     videoEl = document.createElement('video');
     videoEl.style.setProperty('object-fit', 'contain');
     videoEl.id = `video-${participant.sessionId}`;
     videoEl.width = videoDimension.width;
     videoEl.height = videoDimension.height;
-    videoEl.playsInline = true;
-    videoEl.autoplay = true;
-  }
-  if (videoEl.srcObject !== participant.videoStream) {
-    videoEl.srcObject = participant.videoStream || null;
-  }
-  if (
-    !participant.isLocalParticipant &&
-    participant.publishedTracks.includes(SfuModels.TrackType.VIDEO) &&
-    !participant.videoDimension
-  ) {
-    // We need to subscribe to video tracks
-    // We provide the rendered video dimension to save bandwidth
-    call.updateSubscriptionsPartial('video', {
-      [participant.sessionId]: {
-        dimension: {
-          width: videoDimension.width,
-          height: videoDimension.height,
-        },
-      },
-    });
-  }
+    videoEl.dataset.sessionId = participant.sessionId;
 
-  return videoEl;
+    parentContainer.appendChild(videoEl);
+
+    const untrack = call.trackElementVisibility(
+      videoEl,
+      participant.sessionId,
+      'videoTrack',
+    );
+
+    // keep reference to untrack function to call it later
+    videoTrackingCache.set(id, untrack);
+
+    // registers subscription updates and stream changes
+    const unbind = call.bindVideoElement(
+      videoEl,
+      participant.sessionId,
+      'videoTrack',
+    );
+
+    // keep reference to unbind function to call it later
+    videoBindingsCache.set(id, unbind);
+  }
 };
 
-const renderAudio = (participant: StreamVideoParticipant) => {
+const renderAudio = (
+  call: Call,
+  participant: StreamVideoParticipant,
+  parentContainer: HTMLElement,
+) => {
   // We don't render audio for local participant
-  if (participant.isLocalParticipant) {
-    return null;
-  }
+  if (participant.isLocalParticipant) return;
 
-  let audioEl = document.getElementById(
-    `audio-${participant.sessionId}`,
-  ) as HTMLAudioElement | null;
-
+  const id = `audio-${participant.sessionId}`;
+  let audioEl = document.getElementById(id) as HTMLAudioElement | null;
   if (!audioEl) {
     audioEl = document.createElement('audio');
-    audioEl.id = `audio-${participant.sessionId}`;
-    audioEl.autoplay = true;
-  }
+    audioEl.id = id;
+    audioEl.dataset.sessionId = participant.sessionId;
 
-  if (audioEl.srcObject !== participant.audioStream) {
-    audioEl.srcObject = participant.audioStream || null;
-  }
+    parentContainer.appendChild(audioEl);
 
-  return audioEl;
+    // registers subscription updates and stream changes for audio
+    const unbind = call.bindAudioElement(audioEl, participant.sessionId);
+
+    // keep reference to unbind function to call it later
+    audioBindingsCache.set(id, unbind);
+  }
 };
 
 export const renderParticipant = (
   call: Call,
   participant: StreamVideoParticipant,
+  parentContainer: HTMLElement,
 ) => {
-  return {
-    audioEl: renderAudio(participant),
-    videoEl: renderVideo(call, participant),
-  };
+  renderAudio(call, participant, parentContainer);
+  renderVideo(call, participant, parentContainer);
+};
+
+export const cleanupParticipant = (sessionId: string) => {
+  const unbindVideo = videoBindingsCache.get(`video-${sessionId}`);
+  if (unbindVideo) {
+    unbindVideo();
+    videoBindingsCache.delete(`video-${sessionId}`);
+  }
+
+  const untrackVideo = videoTrackingCache.get(`video-${sessionId}`);
+  if (untrackVideo) {
+    untrackVideo();
+    videoTrackingCache.delete(`video-${sessionId}`);
+  }
+
+  const unbindAudio = audioBindingsCache.get(`audio-${sessionId}`);
+  if (unbindAudio) {
+    unbindAudio();
+    audioBindingsCache.delete(`audio-${sessionId}`);
+  }
 };