+++ sample-apps/react/egress-composite/src/components/layouts/DominantSpeaker/AudioTracks.tsx
@@ -10,9 +10,8 @@ export const AudioTracks = (props: {
       {participants.map((participant) => (
         <Audio
           key={participant.sessionId}
-          audioStream={participant.audioStream}
+          participant={participant}
           muted={participant.sessionId === dominantSpeaker?.sessionId}
-          data-user-id={participant.userId}
         />
       ))}
     </div>