+++ sample-apps/react/egress-composite/src/main.tsx
@@ -48,7 +48,7 @@ window.setupLayout = (configuration: ConfigurationValue) => {
 //     call_id: "<call_id>",
 //     layout: "grid",
 //     screenshare_layout: "spotlight",
-
+//     ext_css: "https://github.githubassets.com/assets/light-983b05c0927a.css",
 //     options: {
 //       "logo.image_url": "https://theme.zdassets.com/theme_assets/9442057/efc3820e436f9150bc8cf34267fff4df052a1f9c.png",
 //     }