+++ packages/react-bindings/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+### [0.2.11](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.2.10...@stream-io/video-react-bindings-0.2.11) (2023-08-30)
+
+### Dependency Updates
+
+* `@stream-io/video-client` updated to version `0.1.0`
 ### [0.2.10](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-react-bindings-0.2.9...@stream-io/video-react-bindings-0.2.10) (2023-08-29)
 
 ### Dependency Updates